Put depenency handling back into main CMakeLists
The VCPKG toolchain file must be included before the `project`
command, therefore  we can't have it in the dependencies/CMakeLists.txt.

include_subdirectory is for including subprojects, thus any find_package
or USE_VCPKG command inside there only applies to the subproject, not to
the main project. The idea of the dependencies/ subdirectory is to have
3rdparty dependencies that we include in our source tree directly (e.g.
via git submodules) in there.