Apply the default_init_allocator for NUMA
Using the default_init_allocator doesn't really change anything for the
input vector as long as we don't change the initialization (which this
commit doesn't do), but it changes the way the output vector is
initialized (each mempage is allocated on the NUMA node which first
writes to it). This should not make any difference for fully random
index data, but should make a considerable difference for local scatter
(not tested yet).